[Test] `testrand` package moved to `common` directory
