+++ common/testrand/rand.go
