+++ node/cn/snap/sync_test.go
@@ -38,9 +38,9 @@ import (
 	"github.com/kaiachain/kaia/blockchain/types/account"
 	"github.com/kaiachain/kaia/blockchain/types/accountkey"
 	"github.com/kaiachain/kaia/common"
+	"github.com/kaiachain/kaia/common/testrand"
 	"github.com/kaiachain/kaia/crypto"
 	"github.com/kaiachain/kaia/log"
-	"github.com/kaiachain/kaia/node/cn/snap/testrand"
 	"github.com/kaiachain/kaia/params"
 	"github.com/kaiachain/kaia/rlp"
 	"github.com/kaiachain/kaia/storage/database"
